Partly reworked serial connection
You now have to use the carriage return character (ASCII 13) for the
input to be processed. Also, the Arduino now returns a little welcome
message on first contact with a serial partner and checks for messages
over serial will only be executed if there has already been a serial
connection/a serial connection is currently established.