Added the option to use other serial connections
Set 'SERIALCON' at the top of the program to your desired serial
connection, e.g. 'Serial', 'Serial1', ... .
This is necessary for boards like Arduino Micro or Leonardo, if you want
to use the onboard pins instead of the connection over usb provided by
the bootloader.
Also reworked the section with variables for the end-user; now using
compiler directives.